// PHP // // Process Ajax class : Protection de l'écriture d'un fichier pour qu'il ne soit plus créé alors que la recette n'a pas été insérée proprement dans la database - Fix du répertoire d'image - Le renvoi  du répertoire sera différent dans la database pour permettre une meilleure intégration avec le JavaScript - // CheckInput class : Modification de la fonction checkInputs() pour la rendre plus lisible - Ajout d'un array permettant de filtrer les clés qui peuvent être vides - // Recipe Controller class : Utilisation du paramètre optionnel du constructeur pour l'insertion d'une nouvelle recette dans la DB - Il n'y aura plus d'erreur de recipe_id vide lors de la création de recette - // Recipe class : Protection du code pour ne tenter de supprimer une image que si celle-ci a réellement été trouvée dans la database - // Read : Petit fix lors de l'insertion d'un commentaire - Il n'y aura plus d'erreur lorsqu'aucuns commentaires n'existent pour cette recette -

+++ classes/Process_Ajax.class.php
@@ -52,7 +52,7 @@ class Process_Ajax
                 // 'persons' => json_encode($dataTest)
             ];
             // print_r($this->is_Post);
-            $setRecipe = new RecipeView($this->getDatas);
+            $setRecipe = new RecipeView($this->getDatas, 'creation');
             $this->is_Post ? $recipeId = $setRecipe->insertRecipeTest() : $recipeId = $setRecipe->updateRecipeInfoById();
             $this->send_Status = 'success';
             if (isset($recipeId['status']) && $recipeId['status']['update_status'] === 'RCPUPDTSTMTEXECNT') {
@@ -74,14 +74,20 @@ class Process_Ajax
                     // On vérifie les extensions autorisée
                     $allowedExtensions = ['jpg', 'jpeg', 'gif', 'png'];
                     $new_name = time() . '.' . $extension;
-                    $dir = 'uploads/'. $loggedUser['email'] . '/recipes_images/' . $recipeId;
-                    $file = $dir . '/' . $new_name;
-                    if (in_array($extension, $allowedExtensions)) {
-                        makeDir($dir);
-                        move_uploaded_file($this->Post_Files['file']['tmp_name'], $file);
+                    $database_Dir = 'uploads/'. $loggedUser['email'] . '/recipes_images/' . $recipeId;
+                    $file_In_Database = $database_Dir . '/' . $new_name;
+
+                    $file_Upload_Dir = '../uploads/'. $loggedUser['email'] . '/recipes_images/' . $recipeId . '/';
+                    $file_In_Upload_Dir = $file_Upload_Dir . '/' . $new_name;
+                    // echo($dir);
+                    // echo($fileUploadDir);
+                    if (isset($recipeId) && in_array($extension, $allowedExtensions)) {
+                        makeDir($file_Upload_Dir);
+                        move_uploaded_file($this->Post_Files['file']['tmp_name'], $file_In_Upload_Dir);
                         $this->isImageSent = true;
-                        $image_Data = ['recipeId' => $recipeId, 'fileName' => $new_name, 'filePath' => $file];
+                        $image_Data = ['recipeId' => $recipeId, 'fileName' => $new_name, 'filePath' => $file_In_Database];
                         $setRecipe->deleteImage($recipeId);
+                        // if ($recipeId) $setRecipe->deleteImage($recipeId);
                         $setRecipe->insertImage($image_Data);
                         $this->send_Status = 'success';
                     } else {
